Add today 2021/15 and some readmes
I made a Grid interface, which I might abstract out into
the `aoc` root (especially as Go 1.18 has generics) and
create a simple `ByteGrid` and `IntGrid` implementation.

Turns out today's puzzle needed an abstraction over a simple
grid (to pretend it was 25 times bigger!) - but I don't think that
is common.

I may want a `SparseGrid` as well, that can be used for infinite
planes...